Move refinedBy code into its own file
